Fix docstring example, move test_observe into pyspark.sql.test.test_dataframe
